+++ client/helpers/index.ts
@@ -1,36 +1,6 @@
 import get from 'get-value'
 import set from 'set-value'
 
-/**
- * Converts string to array buffer
- *
- * @param {string} str String
- */
-export function stringToArrayBuffer(str: string): ArrayBuffer {
-  const buf = new ArrayBuffer(str.length)
-  const view = new Uint8Array(buf)
-  for (let i = 0; i !== str.length; ++i) {
-    view[i] = str.charCodeAt(i) & 0xff
-  }
-  return buf
-}
-
-/**
- * Currency display
- *
- * @param {number} num Number
- * @param {string} currency Currency
- * @param {number} minimumFractionDigits Minimum fraction digits
- */
-// eslint-disable-next-line @typescript-eslint/no-inferrable-types
-export function currencyDisplay(num: number, currency: string = 'NOK', minimumFractionDigits: number = 0): string {
-  const formatter = new Intl.NumberFormat('en-US', {
-    style: 'currency',
-    currency,
-    minimumFractionDigits
-  })
-  return formatter.format(num)
-}
 
 /**
  * Get value from object