+++ server/api/graphql/context.ts
@@ -31,9 +31,9 @@ export class Context {
   public container?: ContainerInstance
 
   /**
-   * Is authorized
+   * Permissions
    */
-  public isAuthorized?: boolean
+  public permissions?: string[]
 }
 
 /**
@@ -43,29 +43,22 @@ export class Context {
  */
 export const createContext = async (request: Express.Request): Promise<Context> => {
   try {
-    let isAuthorized = false
-    let user = null
-    let subscription = get(request, 'user.subscription')
+    let context: Context = {}
+    context.userId = null
+    context.subscription = get(request, 'user.subscription')
+    context.permissions = []
     if (!!request.token) {
-      subscription = await new SubscriptionService().findSubscriptionWithToken(request.token)
-      if (!subscription) throw new AuthenticationError(null)
-      isAuthorized = true
+      const token = await new SubscriptionService().getToken(request.token)
+      if (!token) throw new AuthenticationError('Token is invalid.')
+      context = { ...context, ...token }
     } else {
-      isAuthorized = !!get(request, 'user')
-      user = isAuthorized && { id: request.user.id }
+      context.userId = !!context.permissions && get(request, 'user.id')
     }
-    const requestId = Math.floor(Math.random() * Number.MAX_SAFE_INTEGER).toString()
-    const container = Container.of(requestId)
-    const context: Context = {
-      container,
-      subscription,
-      requestId,
-      userId: user.id,
-      isAuthorized
-    }
-    container.set({ id: 'CONTEXT', transient: true, value: context })
-    container.set({ id: 'REQUEST', transient: true, value: request })
-    debug(`Creating context for request ${requestId}`)
+    context.requestId = Math.floor(Math.random() * Number.MAX_SAFE_INTEGER).toString()
+    context.container = Container.of(context.requestId)
+    context.container.set({ id: 'CONTEXT', transient: true, value: context })
+    context.container.set({ id: 'REQUEST', transient: true, value: request })
+    debug(`Creating context for request ${context.requestId}`)
     return context
   } catch (error) {
     throw error