+++ server/api/graphql/resolvers/apiToken.types.ts
@@ -1,6 +1,6 @@
 /* eslint-disable max-classes-per-file */
 import 'reflect-metadata'
-import { Field, ID, ObjectType } from 'type-graphql'
+import { Field, ID, InputType, ObjectType } from 'type-graphql'
 import { simpleResolvers } from '../config'
 
 @ObjectType({ description: 'A type that describes a ApiToken', simpleResolvers: simpleResolvers.ApiToken })
@@ -10,4 +10,19 @@ export class ApiToken {
 
   @Field()
   created: Date
+
+  @Field()
+  expires: string
+}
+
+@InputType({ description: 'Input object for ApiToken used in mutation addApiToken' })
+export class ApiTokenInput {
+  @Field(() => ID)
+  name?: string
+
+  @Field()
+  expires?: string
+
+  @Field(() => [String])
+  permissions?: string[]
 }