+++ server/api/graphql/resolvers/notification.forecast.ts
@@ -14,25 +14,25 @@ import { AzStorageService } from '../../services'
  * * forecast.enabled
  * * forecast.notifications
  *
- * @param ctx
- * @param azstorage
- * @param template
- * @param locale
+ * @param {Context} ctx Context
+ * @param {AzStorageService} azstorage Azure Table Storage service
+ * @param {string} template Template
+ * @param {string} locale Locale
  */
 export default async function (ctx: Context, azstorage: AzStorageService, template: string, locale: string) {
-  if (!get(ctx, 'user.subscription.settings.forecast.enabled', { default: false })) return []
+  if (!get(ctx, 'subscription.settings.forecast.enabled', { default: false })) return []
   const currentWeek = utils.getWeek()
   const periods = []
   const unforecastedPeriods = []
 
-  for (let i = 1; i <= get(ctx, 'user.subscription.settings.forecast.notifications', { default: 2 }); i++) {
+  for (let i = 1; i <= get(ctx, 'subscription.settings.forecast.notifications', { default: 2 }); i++) {
     periods.push(...getPeriods(utils.startOfWeek(currentWeek + i), utils.endOfWeek(currentWeek + i), locale))
   }
 
-  const forecastedPeriods = (await azstorage.getForecastedPeriods({
+  const forecastedPeriods = await azstorage.getForecastedPeriods({
     resourceId: ctx.userId,
     year: utils.getYear()
-  })) as any[]
+  })
 
   periods.forEach((period) => {
     if (!find(forecastedPeriods, (cp) => cp.periodId === period.id)) unforecastedPeriods.push(period)