+++ server/api/graphql/resolvers/notification.resolver.ts
@@ -35,7 +35,6 @@ export class NotificationResolver {
     @Arg('locale') locale: string,
     @Ctx() ctx: Context
   ) {
-    if (!ctx.userId) return { success: false, error: null }
     const notifications = await Promise.all([
       unconfirmedPeriods(ctx, this._azstorage, templates.unconfirmedPeriods, locale),
       forecast(ctx, this._azstorage, templates.forecast, locale)