+++ server/api/graphql/resolvers/role.resolver.ts
@@ -3,6 +3,7 @@ import { Arg, Authorized, Mutation, Query, Resolver } from 'type-graphql'
 import { Service } from 'typedi'
 import { pick } from 'underscore'
 import { AzStorageService } from '../../services'
+import { IAuthOptions } from '../authChecker'
 import { Role, RoleInput } from './role.types'
 import { BaseResult } from './types'
 
@@ -16,12 +17,12 @@ export class RoleResolver {
    *
    * @param {AzStorageService} _azstorage AzStorageService
    */
-  constructor(private readonly _azstorage: AzStorageService) {}
+  constructor(private readonly _azstorage: AzStorageService) { }
 
   /**
    * Get roles
    */
-  @Authorized()
+  @Authorized<IAuthOptions>({ userContext: true })
   @Query(() => [Role], { description: 'Get roles' })
   async roles() {
     return await this._azstorage.getRoles()
@@ -29,11 +30,13 @@ export class RoleResolver {
 
   /**
    * Add or update role
+   * 
+   * @permission MANAGE_ROLESPERMISSIONS (cd52a735)
    *
    * @param {RoleInput} role Role
    * @param {boolean} update Update
    */
-  @Authorized()
+  @Authorized<IAuthOptions>({ permission: 'cd52a735' })
   @Mutation(() => BaseResult, { description: 'Add or update role' })
   async addOrUpdateRole(
     @Arg('role', () => RoleInput) role: RoleInput,