+++ server/api/graphql/resolvers/subscription.resolver.ts
@@ -3,6 +3,7 @@ import { Arg, Authorized, Ctx, Mutation, Query, Resolver } from 'type-graphql'
 import { Service } from 'typedi'
 import { pick } from 'underscore'
 import { SubscriptionService } from '../../services'
+import { IAuthOptions } from '../authChecker'
 import { Context } from '../context'
 import { Subscription, SubscriptionSettingsInput } from './subscription.types'
 import { BaseResult } from './types'
@@ -17,12 +18,12 @@ export class SubscriptionResolver {
    *
    * @param {SubscriptionService} _subscription SubscriptionService
    */
-  constructor(private readonly _subscription: SubscriptionService) {}
+  constructor(private readonly _subscription: SubscriptionService) { }
 
   /**
    * Get current subscription
    */
-  @Authorized()
+  @Authorized({ userContext: true })
   @Query(() => Subscription, { description: 'Get current subscription' })
   async subscription(@Ctx() ctx: Context): Promise<Subscription> {
     try {
@@ -34,11 +35,13 @@ export class SubscriptionResolver {
 
   /**
    * Update subscription
+   * 
+   * @permission MANAGE_SUBSCRIPTION (67ba6efc)
    *
    * @param {string} id Subscription ID
    * @param {SubscriptionSettingsInput} settings Settings
    */
-  @Authorized()
+  @Authorized<IAuthOptions>({ permission: '67ba6efc' })
   @Mutation(() => BaseResult, { description: 'Update subscription' })
   async updateSubscription(
     @Arg('id') id: string,